Add smoke test of joins option.
This doesn't _really_ test the join option but off the top of my head
I'm not sure what the best way to do so is. If there's a method that
stops ActiveRecord from fetching association info from the database that
would probably work but I don't know it. This at least test that
specifying a join doesn't break anything else.